boq-agent/pyproject.toml
[project]
name = "boq-agent"
version = "0.1.0"
description = "AI-powered Bill of Quantities extraction from tender documents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "H2E Team" }
]
keywords = ["boq", "construction", "tender", "document-extraction", "ocr", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
]

dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # PDF parsing
    "pdfplumber>=0.10.0",
    "pypdf>=4.0.0",
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",

    # Document parsing
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "xlrd>=2.0.0",

    # LLM & NLP
    "openai>=1.12.0",
    "tiktoken>=0.6.0",

    # Config & utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
boq-agent = "boq_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/boq_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["boq_agent"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
